* commit '8bae73bfff05e2b55cf334b5a9005cb602924490':
  PWP-59 Update MQ configuration messages

+++ powerimo-sqb-core/src/main/java/org/powerimo/sqb/SearchParamsProvider.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface SearchParamsProvider {
      List<Condition> getConditions();
      String getOrderBy();
-     Integer getLimit();
-     Integer getLimitOffset();
+     Long getLimit();
+     Long getLimitOffset();
 }
